requirements.txt
anyio==3.6.2
asgiref==3.7.2
beautifulsoup4==4.12.2
cachetools==5.2.0
certifi==2021.10.8
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.6
cloudinary==1.30.0
colorama==0.4.6
confusable-homoglyphs==3.2.0
coreapi==2.3.3
coreapi-cli==1.0.9
coreschema==0.0.4
cryptography==37.0.4
cssselect==1.2.0
cssutils==2.6.0
defusedxml==0.7.1
dj-database-url==2.0.0
Django==4.2.2
django-appconf==1.0.5
django-bootstrap4==23.1
django-cors-headers==3.13.0
django-crispy-forms==1.14.0
django-imagekit==4.1.0
django-on-heroku==1.1.2
django-registration==3.3
django-simple-mail==2.3.5
django-tinymce==3.4.0
django-utils-six==2.0
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-pagination==1.1
djangorestframework-simplejwt==5.2.2
djangorestframework-xml==2.0.0
dnspython==2.2.1
drf-yasg==1.21.7
email-validator==1.3.0
gunicorn==20.1.0
h11==0.14.0
html2text==2020.1.16
httpcore==0.16.3
httpx==0.23.1
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
lxml==4.9.2
MarkupSafe==2.1.3
packaging==23.1
pilkit==2.0
Pillow==9.3.0
premailer==3.10.0
psycopg2-binary==2.9.5
pycodestyle==2.7.0
pycparser==2.21
pydantic==1.10.2
PyJWT==1.7.1
python-dateutil==2.8.2
python-decouple==3.5
pytz==2022.6
pyuploadcare==3.1.0
PyYAML==6.0.1
requests==2.28.1
rfc3986==1.5.0
ruamel.yaml==0.17.32
ruamel.yaml.clib==0.2.7
six==1.16.0
sniffio==1.3.0
soupsieve==2.2.1
sqlparse==0.4.2
toml==0.10.2
typing_extensions==4.4.0
tzdata==2022.7
uritemplate==4.1.1
urllib3==1.26.7
whitenoise==6.2.0
zipp==3.6.0
